finalizando estrutura do modulo de estimação de pose / corrrigindo detalhes
