--- src/ml/new_pre_process.py
@@ -1,1 +1,0 @@
-import numpya as np
